// util/types.ts
export type lettresType = {
    Num_id: string;
    lettre_id: string;
    texte: string;
  };
  
 export type lettresHebreuType = {
    Igroys_address: string;
    Igroys_bsd: string;
    Igroys_city: string;
    Igroys_comment: string;
    Igroys_signature: string;
    Igroys_text: string;
    Page_number: string;
    book_page: string;
    idx: number;
    summary: string;
    vol: number;
  };